Migrate Decompose* to utils.  Migrate and export UnqualifiedName(Var).  Correct nil-deref issue.
